+++ tox.ini
@@ -13,5 +13,4 @@ usedevelop=true
 deps = mypy>=1.7
 commands = mypy -p navmazing
 [flake8]
-max_line_length = 100
-ignore = E128,E811
+max_line_length = 88
